Crate hebrides

Implementations for real and complex numbers.

Hebrides provides two main structs: Real and Complex, which together form the basis of the Marlin ecocystem, which as a whole constitutes the whole of the implementation of DOSs in Rust.

Structs

Angle

Representation of angular values.

Complex

Representation of complex numbers

ConversionError

Error type for errors involving conversion between values.

DomainError

Error type for errors involving domain restrictions.

Real

Representation of real numbers.