Struct hebrides::Complex

pub struct Complex { /* private fields */ }

Representation of complex numbers

Implementations

impl Complex

pub const ZERO: Complex = _

shortcut for encoding 0.0 as Complex::new(Real::ZERO, Real::ZERO)

pub const ONE: Complex = _

shortcut for encoding 1.0 as Complex::new(Real::ONE, Real::ZERO)

pub const I: Complex = _

shortcut for encoding the imaginary unit as Complex::new(Real::ZERO, Real::ONE)

pub fn new(real: f64, imag: f64) -> Self

Constructs a Complex from two f64 values

pub fn is_real(&self) -> bool

Returns whether or not self is real.

let z = Complex::new(3.0, 2.0);
assert!(!z.is_real());
let x = Complex::new(1.0, 0.0);
assert!(x.is_real());
let y = Complex::new(0.0, 1.0);
assert!(!y.is_real());

pub fn is_imaginary(&self) -> bool

Returns whether or not self is imaginary.

let z = Complex::new(3.0, 2.0);
assert!(!z.is_imaginary());
let x = Complex::new(23.0, 0.0);
assert!(!x.is_imaginary());
let y = Complex::new(0.0, 12.0);
assert!(y.is_imaginary());

pub fn abs(&self) -> Complex

Absolute value (aka. complex norm).

let z = Complex::new(3.0, 4.0);
assert_eq!(z.abs(), Complex::new(5.0, 0.0));

pub fn exp(&self) -> Complex

Complex exponentiation.

Evaluates e to the power of self.

let e = Complex::new(std::f64::consts::E, 0.0);
let z = Complex::new(1.0, std::f64::consts::FRAC_PI_2);
assert_eq!(z.exp(), e*Complex::I);
let q = Complex::new(1.0, 1.0);
assert_eq!(q.exp(), Complex::new(1.4686939399158851, 2.2873552871788423));
let w = Complex::new(3.0, 7.0);
assert_eq!(w.exp(), Complex::new(15.142531566086864, 13.195928586605717))

pub fn sin(&self) -> Complex

Complex sine.

let z = Complex::new(1.0, std::f64::consts::FRAC_PI_2);
assert_eq!(z.sin(), Complex::new(2.1114008854951747, 1.2433971034084503));

pub fn cos(&self) -> Complex

Complex cosine.

let z = Complex::new(0.0, 0.0);
assert_eq!(z.cos(), Complex::new(1.0, 0.0));
let w = Complex::new(1.0, 2.0);
assert_eq!(w.cos(), Complex::new(2.0327230070196656, -3.0518977991517997));

pub fn tan(&self) -> Complex

Complex tangent.

let z = Complex::new(2.0, 3.0);
assert_eq!(z.tan(), Complex::new(-0.0037640256415040815, 1.0032386273536098));

pub fn arcsin(&self) -> Complex

Complex arcsin.

Even if self is real, this will return only a Complex and not an Angle as might be expected based off of the behavior of the arcsine implementation on Real.

Mathematical justification for the implementation can be found at ProofWiki.

let x = Complex::new(0.8, 0.0);
assert_eq!(x.arcsin(), Real::new(Real::new(0.8).arcsin().unwrap().to_radians()).to_complex());
let z = Complex::new(2.0, 3.0);
assert_eq!(z.arcsin(), Complex::new(1.5413069599437077, -1.3679718444908728));

pub fn arccos(&self) -> Complex

Complex arccos.

Even if self is real, this will return only a Complex and not an Angle as might be expected based off of the behavior of the arccosine implementation on Real.

Mathematical justification for the implementation can be found at ProofWiki.

let x = Complex::new(0.8, 0.0);
assert_eq!(x.arccos(), Real::new(Real::new(0.8).arccos().unwrap().to_radians()).to_complex());
let z = Complex::new(2.0, 3.0);
assert_eq!(z.arccos(), Complex::new(1.0001435424737972, -1.9833870299165355));

pub fn arctan(&self) -> Complex

Complex arctan.

Even if self is real, this will return only a Complex and not an Angle as might be expected based off of the behavior of the arctangent implementation on Real.

Mathematical justification for the implementation can be found at ProofWiki.

let x = Complex::new(0.8, 0.0);
assert_eq!(x.arctan(), Real::new(Real::new(0.8).arctan().to_radians()).to_complex());
let z = Complex::new(2.0, 3.0);
assert_eq!(z.arctan(), Complex::new(1.4099210495965755, 0.22907268296853875));

pub fn sinh(&self) -> Complex

Complex hyperbolic sine.

let z = Complex::new(3.0, 2.0);
assert_eq!(z.sinh(), Complex::new(-4.168906959966565, 9.154499146911428));

pub fn cosh(&self) -> Complex

Complex hyperbolic cosine.

let z = Complex::new(3.0, 2.0);
assert_eq!(z.cosh(), Complex::new(-4.189625690968807, 9.109227893755337));

pub fn tanh(&self) -> Complex

Complex hyperbolic tangent.

let z = Complex::new(3.0, 2.0);
assert_eq!(z.tanh(), Complex::new(1.0032386273536098, -0.0037640256415040815));

pub fn arcsinh(&self) -> Complex

Complex inverse hyperbolic sine.

pub fn arccosh(&self) -> Complex

Complex inverse hyperbolic cosine.

pub fn arctanh(&self) -> Complex

Complex inverse hyperbolic tangent.

pub fn ln(&self) -> Complex

Complex natural logarithm.

Computes the natural logarithm of self on the principal branch of Ln(x).

let z = Complex::new(3.0, 4.0);
assert_eq!(z.ln(), Complex::new(5.0_f64.ln(), 0.8_f64.asin()));

pub fn norm(&self) -> Real

Complex norm.

Returns the absolute value of self. If self is interpreted as a vector, this is its length. The norm is always a positive real number.

let z = Complex::new(3.0, 4.0);
assert_eq!(z.norm(), Real::new(5.0));
let w = Complex::new(5.0, 12.0);
assert_eq!(w.norm(), Real::new(13.0));

pub fn azimuthal(&self) -> Angle

Azimuthal angle.

Returns the azimuthal angle of self. Its range is from [0, 2pi).

let w = Complex::new(3.0, 4.0);
assert_eq!(w.azimuthal().to_radians(), 0.8_f64.asin());
let v = Complex::new(-3.0, 4.0);
assert_eq!(v.azimuthal().to_radians(), std::f64::consts::PI - 0.8_f64.asin());

pub fn squared(&self) -> Complex

Returns self squared.

let z = Complex::new(-2.0, 3.0);
assert_eq!(z.squared(), Complex::new(-5.0, -12.0));

pub fn sqrt(&self) -> Complex

Returns the square root of self.

let z = Complex::new(2.0, 3.0);
assert_eq!(z.sqrt(), Complex::new(1.67414922803554, 0.8959774761298381));