Struct hebrides::Real

pub struct Real { /* private fields */ }

Representation of real numbers.

Real numbers can be thought of, without loss of generality, as the set of integers union the set of rationals union the set of irrationals. The only primitive data type native to Rust that is able to represent all three of these classes is f64 (not counting the impossibility of completely describing irrational numbers as decimals), which is why it’s used as the struct’s internal representation of real values.

Implementations

impl Real

pub const ONE: Real = _

shortcut for describing Real::new(0.0)

pub const ZERO: Real = _

shortcut for describing Real::new(1.0)

pub fn new(inner: f64) -> Self

Constructs a Real from an f64.

pub fn value(&self) -> f64

Returns a copy of self as an f64.

let y = Real::new(12.34);
assert_eq!(y.value(), 12.34);

pub fn positive(&self) -> bool

Whether or not self is positive.

let x = Real::new(0.4);
assert!(x.positive());
let y = Real::new(-0.4);
assert!(!y.positive());

pub fn negative(&self) -> bool

Whether or not self is negative.

let x = Real::new(0.5);
assert!(!x.negative());
let y = Real::new(-0.5);
assert!(y.negative());

pub fn abs(&self) -> Self

Absolute value of self.

let x = Real::ONE;
assert_eq!(x.abs(), Real::ONE);
let y = -Real::ONE;
assert_eq!(y.abs(), Real::ONE);

pub fn ceil(&self) -> Self

Returns the ceiling of self.

let x = Real::new(0.02342);
assert_eq!(x.ceil(), Real::ONE);

pub fn floor(&self) -> Self

Returns the floor of self.

let x = Real::new(0.02322);
assert_eq!(x.floor(), Real::ZERO);

pub fn sin(&self) -> Self

Sine.

let theta = Real::new(std::f64::consts::PI);
assert_eq!(theta.sin(), Real::ZERO);

pub fn cos(&self) -> Self

Cosine.

let theta = Real::new(std::f64::consts::PI);
assert_eq!(theta.cos(), -Real::ONE);

pub fn tan(&self) -> Self

Tangent.

let theta = Real::new(std::f64::consts::FRAC_PI_4);
assert_eq!(theta.tan(), Real::ONE);

pub fn arcsin(&self) -> Result<Angle, DomainError>

Arcsine.

Inputs to arcsin must be between negative one and one; its range extends from [pi/2, -pi/2].

let h = Real::new(4.0 / 5.0);
assert_eq!(h.arcsin().unwrap().to_radians(), 0.8_f64.asin());

pub fn arccos(&self) -> Result<Angle, DomainError>

Arccosine.

Inputs to arccos must be between negative one and one; its range extends from [0, pi].

let h = Real::new(3.0 / 5.0);
assert_eq!(h.arccos().unwrap().to_radians(), 0.6_f64.acos());

pub fn arctan(&self) -> Angle

Arctangent.

The range of arctan extends between [-pi/2, pi/2].

let h = Real::ONE;
assert_eq!(h.arctan().to_radians(), 1.0_f64.atan());

pub fn sinh(&self) -> Self

Hyperbolic sine.

let x = Real::new(2.0);
assert_eq!(x.sinh(), Real::new(2.0_f64.sinh()));

pub fn cosh(&self) -> Self

Hyperbolic cosine.

let x = Real::new(3.0);
assert_eq!(x.cosh(), Real::new(3.0_f64.cosh()));

pub fn tanh(&self) -> Self

Hyperbolic tangent.

let x = Real::new(4.0);
assert_eq!(x.tanh(), Real::new(4.0_f64.tanh()));

pub fn arcsinh(&self) -> Angle

Hyperbolic arcsine.

let x = Real::new(3.0);
assert_eq!(x.arcsinh().to_radians(), 3.0_f64.asinh());

pub fn arccosh(&self) -> Result<Angle, DomainError>

Hyperbolic arccosine.

Inputs to arccosh must be between [1, infinity].

let x = Real::new(2.0);
assert_eq!(x.arccosh().unwrap().to_radians(), 2.0_f64.acosh());

pub fn arctanh(&self) -> Result<Angle, DomainError>

Hyperbolic arctangent.

Inputs to arctanh must be between [-1, 1].

let x = Real::new(0.5);
assert_eq!(x.arctanh().unwrap().to_radians(), 0.5_f64.atanh());

pub fn exp(&self) -> Self

e to the power of self.

let x = Real::ZERO;
assert_eq!(x.exp(), Real::ONE);
let y = Real::ONE;
assert_eq!(Real::ONE.exp(), Real::new(std::f64::consts::E));

pub fn pow(&self, power: Self) -> Self

self to the power of power.

let x = Real::new(2.0);
assert_eq!(x, x.pow(Real::ONE));
assert_eq!(x*x, x.pow(Real::new(2.0)));

pub fn powf(&self, power: f64) -> Result<Self, DomainError>

self to the power of power.

Only accepts non-negative arguments; imaginary values are not included in the output of powf.

let x = Real::new(2.0);
assert_eq!(x, x.powf(1.0).unwrap());
assert_eq!(x*x, x.powf(2.0).unwrap());

pub fn powi(&self, power: i64) -> Self

self to the power of power.

let x = Real::new(2.0);
assert_eq!(x, x.powi(1));
assert_eq!(x*x, x.powi(2));

pub fn squared(&self) -> Self

Returns the square of self.

let x = Real::new(2.0);
assert_eq!(x.squared(), Real::new(4.0));
let y = Real::new(-2.0);
assert_eq!(y.squared(), Real::new(4.0));

pub fn sqrt(&self) -> Result<Self, DomainError>

Returns the square root of self.

let x = Real::new(4.0);
assert_eq!(x.sqrt().unwrap(), Real::new(2.0));

pub fn ln(&self) -> Result<Self, DomainError>

Natural logarithm.

Inputs to ln must be greater than zero.

let e = Real::new(std::f64::consts::E);
assert_eq!(e.ln().unwrap(), Real::ONE);
let x = Real::ONE;
assert_eq!(x.ln().unwrap(), Real::new(1.0_f64.ln()));

pub fn log(&self, base: Real) -> Result<Self, DomainError>

Logarithm with base ‘base’.

Inputs for base must be greater than zero, and as for all logarithms, self must also be greater than zero.

let x = Real::new(5.0);
assert_eq!(x.log(Real::new(5.0)).unwrap(), Real::ONE);

pub fn logf(&self, base: f64) -> Result<Self, DomainError>

Logarithm with base ‘base’.

Inputs for base must be greater than zero, and as for all logarithms, self must also be greater than zero.

let x = Real::new(5.0);
assert_eq!(x.logf(5.0).unwrap(), Real::ONE);

pub fn logi(&self, base: i64) -> Result<Self, DomainError>

Logarithm with base ‘base’.

Inputs for base must be greater than zero, and as for all logarithms, self must also be greater than zero.

let x = Real::new(5.0);
assert_eq!(x.logi(5).unwrap(), Real::ONE);

pub fn to_complex(&self) -> Complex

Constructs a Complex from self.

let x = Real::new(3.0);
assert_eq!(x.to_complex(), Complex::new(3.0, 0.0));